Experimental: Greater size for variables to store data

+++ src/experimental/Input.kt
@@ -2,7 +2,7 @@ package experimental
 
 import javax.swing.JOptionPane
 
-fun readInput(variable: String): Int {
+fun readInput(variable: String): Long {
     val inputString = JOptionPane.showInputDialog(null, "Variable: $variable")
-    return inputString.toInt()
+    return inputString.toLong()
 }
