+++ src/experimental/Output.kt
@@ -2,7 +2,7 @@ package experimental
 
 import javax.swing.JOptionPane
 
-fun showOutput(value: Int?) {
+fun showOutput(value: Long?) {
     var output = "Output: $value"
     JOptionPane.showMessageDialog(null, output, "", JOptionPane.INFORMATION_MESSAGE)
 }
